create

+++ scripts/merge.sh
@@ -1,6 +1,6 @@
 export $(grep -v '^#' settings.env | xargs -d '\n')
 
-read -p "This will copy changes from repo to prod, type Y to continue: " answer
+read -p "This will copy changes from repo to ${prod_instance_name}, type Y to continue: " answer
 if [ "${answer}" != "Y" ]; then
     echo "OK, quitting."
     exit
@@ -25,15 +25,31 @@ wget $prod_db_wallet_preauth -O wallet.zip
 
 echo "*** COPY REPO MASTER TO ${name} ***"
 
-printf "set cloudconfig ./wallet.zip\nconn ${schema}/${pwd}@${name}_high\nlb update -changelog controller.xml\nlb update -changelog data.xml\ntables\nexit" > upd.sql
+printf "set cloudconfig ./wallet.zip\nconn ${schema}/${pwd}@${name}_high\n" > upd.sql
+
+if [ -f "controller.xml" ]; then
+   printf "lb update -changelog controller.xml\n" >> upd.sql
+else
+    echo "controller.xml not found. Not copied to ${prod_instance_name}." >> upd.sql
+fi
+
+for filename in /data/*.xml; do
+    printf "lb update -changelog ${filename}\ntables\nexit" >> upd.sql
+done
+
+cat upd.sql
 
 sql /nolog @./upd.sql
 rm -f upd.sql
 
 if [ -n "${application_id}" ]; then
-    printf "set cloudconfig ./wallet.zip\nconn ${schema}/${pwd}@${name}_high\n@privileges.sql\nlb update -changelog f${application_id}.xml\nexit" > upd_apex.sql
-    sql /nolog @./upd_apex.sql
-    rm -f upd_apex.sql
+    if [ -f "f${application_id}.xml" ]; then
+        printf "set cloudconfig ./wallet.zip\nconn ${schema}/${pwd}@${name}_high\nlb update -changelog f${application_id}.xml\nexit" > upd_apex.sql
+        sql /nolog @./upd_apex.sql
+        rm -f upd_apex.sql
+    else
+        echo "f${application_id}.xml not found. App not copied to ${prod_instance_name}."
+    fi
 fi
 
 rm -f wallet.zip
